Decoupled middleware functions from index.ts
