+++ Back-End/index.ts
@@ -1,12 +1,17 @@
-import express, { Request, Response, NextFunction } from "express";
+import express from "express";
 import cors from "cors";
+import corsOptions from "./middleware/corsMiddleware";
 import dotenv from "dotenv";
 import cookieParser from "cookie-parser";
-import createError from "http-errors";
 import Database from "@controllers/DBController/DBController";
 import HTTP from "@Util/HTTPCodes";
-import rateLimit from "express-rate-limit";
-import * as Sentry from "@sentry/react";
+import {
+	forgotPasswordLimiter,
+	resetPasswordLimiter,
+	loginLimiter,
+	signupLimiter,
+} from "./middleware/rateLimiter";
+import { notFoundHandler, errorHandler } from "./middleware/errorHandler";
 
 //Routes import
 import authRouter from "@routes/auth";
@@ -21,13 +26,6 @@ import Admin from "@routes/adminRoutes";
 import requisiteRouter from "@routes/requisite";
 import feedbackRouter from "@routes/feedback";
 
-// Sentry Backend INIT
-Sentry.init({
-	dsn: process.env.REACT_APP_SENTRY_DSN,
-	integrations: [Sentry.browserTracingIntegration()],
-	tracesSampleRate: 1.0,
-});
-
 //Dev Consts
 const HOPPSCOTCH = "chrome-extension://amknoiejhlmhancpahfcfcfhllgkpbld";
 
@@ -36,37 +34,6 @@ dotenv.config(); //Load environment variables from .env file
 const app = express();
 const PORT = process.env.PORT || 8000;
 const CLIENT = process.env.CLIENT || "http://localhost:3000";
-
-const corsOptions: cors.CorsOptions = {
-	origin: (
-		origin: string | undefined,
-		callback: (err: Error | null, allow?: boolean) => void
-	) => {
-		const allowedOrigins = [
-			"http://localhost:3000",
-			"http://167.71.165.174:3000",
-			"http://159.65.216.141:3000",
-			"chrome-extension://amknoiejhlmhancpahfcfcfhllgkpbld",
-			"https://hoppscotch.io",
-		];
-
-		// Allow requests with no origin (like Postman or server-to-server calls)
-		if (!origin || allowedOrigins.includes(origin)) {
-			callback(null, true); // Allow
-		} else {
-			callback(new Error("Not allowed by CORS")); // Deny
-		}
-	},
-	credentials: true, // Allow cookies to be sent
-	methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
-	allowedHeaders: [
-		"Origin",
-		"X-Requested-With",
-		"Content-Type",
-		"Accept",
-		"Authorization",
-	],
-};
 
 // Apply the CORS middleware
 app.use(cors(corsOptions));
@@ -78,42 +45,6 @@ app.options("*", cors(corsOptions));
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 app.use(cookieParser());
-
-// Rate limiter for "Forgot Password"
-const forgotPasswordLimiter = rateLimit({
-	windowMs: 2 * 60 * 1000, // window should be as large as OTP expiry time
-	max: 5, // Limit each IP to 5 requests every 2 minutes
-	message: { error: "Too many password reset attempts. Try again later." },
-	standardHeaders: true, // Send rate limit info in headers
-	legacyHeaders: false, // Disable X-RateLimit legacy headers
-});
-
-// Rate limiter for "Reset Password"
-const resetPasswordLimiter = rateLimit({
-	windowMs: 2 * 60 * 1000,
-	max: 3, // Limit each IP to 3 reset attempts per window
-	message: { error: "Too many reset attempts. Try again later." },
-	standardHeaders: true,
-	legacyHeaders: false,
-});
-
-// Rate limiter for Login
-const loginLimiter = rateLimit({
-	windowMs: 1 * 60 * 1000, // 1 min window
-	max: 5, // Attempt to login in 5 times a minute
-	message: { error: "Too many password reset attempts. Try again later." },
-	standardHeaders: true,
-	legacyHeaders: false,
-});
-
-// Rate limiter for signup
-const signupLimiter = rateLimit({
-	windowMs: 1 * 60 * 1000,
-	max: 5,
-	message: { error: "Too many password reset attempts. Try again later." },
-	standardHeaders: true,
-	legacyHeaders: false,
-});
 
 // Apply rate limiters of forgot-password and reset-password routes
 app.use("/auth/forgot-password", forgotPasswordLimiter);
@@ -159,9 +90,8 @@ app.get("/test-db", async (req, res) => {
 });
 
 //Handle 404
-app.use((req: Request, res: Response, next: NextFunction) => {
-	next(createError(HTTP.NOT_FOUND, "Page not found!!!"));
-});
+app.use(notFoundHandler);
+app.use(errorHandler);
 
 //Listen for requests
 app.listen(PORT, () => {
